move helper function to service test to use in multiple construction test files
